requirements.txt
molo.core==2.6.13
molo.profiles==0.2.5
molo.commenting==0.3.2
molo.yourwords==0.0.3
molo.servicedirectory==2.1.0
django-extensions
django-google-analytics-app
django-celery
redis
elasticsearch
